backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
qdrant-client>=1.7.0
redis>=5.0.0
PyMuPDF>=1.23.0
llama-index>=0.10.0
llama-index-llms-gemini>=0.1.0
llama-index-embeddings-gemini>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
spacy>=3.7.0
google-generativeai>=0.4.0
python-multipart>=0.0.9